basic error and board checking
added some basic error and board checking to check if user input is
legal and if a move can be made. again, im not able to do much of this
on my own. reading off the instructions and trying to build my own is
the extent of my abilities at this time. with more time maybe i can
start breaking it and rebuilding my own version of it. for now i just
gotta figure out how it works